// frontend/src/api/models/TaskDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Tracker API
 * Task Tracker API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: trnka.roman13@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProjectResponse } from './ProjectResponse';
import {
    ProjectResponseFromJSON,
    ProjectResponseFromJSONTyped,
    ProjectResponseToJSON,
} from './ProjectResponse';
import type { UserDTO } from './UserDTO';
import {
    UserDTOFromJSON,
    UserDTOFromJSONTyped,
    UserDTOToJSON,
} from './UserDTO';

/**
 * 
 * @export
 * @interface TaskDTO
 */
export interface TaskDTO {
    /**
     * 
     * @type {string}
     * @memberof TaskDTO
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskDTO
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskDTO
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof TaskDTO
     */
    dueDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof TaskDTO
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskDTO
     */
    priority?: string;
    /**
     * 
     * @type {UserDTO}
     * @memberof TaskDTO
     */
    user?: UserDTO;
    /**
     * 
     * @type {ProjectResponse}
     * @memberof TaskDTO
     */
    project?: ProjectResponse;
}

/**
 * Check if a given object implements the TaskDTO interface.
 */
export function instanceOfTaskDTO(value: object): value is TaskDTO {
    return true;
}

export function TaskDTOFromJSON(json: any): TaskDTO {
    return TaskDTOFromJSONTyped(json, false);
}

export function TaskDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): TaskDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'dueDate': json['dueDate'] == null ? undefined : (new Date(json['dueDate'])),
        'status': json['status'] == null ? undefined : json['status'],
        'priority': json['priority'] == null ? undefined : json['priority'],
        'user': json['user'] == null ? undefined : UserDTOFromJSON(json['user']),
        'project': json['project'] == null ? undefined : ProjectResponseFromJSON(json['project']),
    };
}

export function TaskDTOToJSON(value?: TaskDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'dueDate': value['dueDate'] == null ? undefined : ((value['dueDate']).toISOString()),
        'status': value['status'],
        'priority': value['priority'],
        'user': UserDTOToJSON(value['user']),
        'project': ProjectResponseToJSON(value['project']),
    };
}